+++ src/Sulu/Bundle/SearchBundle/Search/Configuration/IndexConfigurationProvider.php
@@ -29,7 +29,8 @@ class IndexConfigurationProvider implements IndexConfigurationProviderInterface
             $this->indexConfigurations[$indexName] = new IndexConfiguration(
                 $indexName,
                 isset($indexConfiguration['name']) ? $indexConfiguration['name'] : null,
-                isset($indexConfiguration['security_context']) ? $indexConfiguration['security_context'] : null
+                isset($indexConfiguration['security_context']) ? $indexConfiguration['security_context'] : null,
+                isset($indexConfiguration['contexts']) ? $indexConfiguration['contexts'] : []
             );
         }
     }