+++ CHANGELOG.md
@@ -3,6 +3,9 @@ CHANGELOG for Sulu
 
 * dev-develop
     * BUGFIX      #2104 [ContentBundle]       Show Webspace node on 'copy' and 'move' overlays  
+    * ENHANCEMENT #2520 [ContentBundle]       Delete routes using the DocumentManager
+    * BUGFIX      #2523 [SecurityBundle]      Fixed error with non-visible permission types in matrix
+    * ENHANCEMENT #2522 [All]                 Use correct default phpcr session
     * ENHANCEMENT #2518 [ContentBundle]       Moved parent from BasePageDocument to PageDocument
     * ENHANCEMENT #2507 [SearchBundle]        Changed search adapter to fit new features of MassiveSearchBundle (limit + offset)
     * ENHANCEMENT #2508 [DocumentManager]     Set default structure-type if non given